+++ pytsite/__init__.py
@@ -104,7 +104,7 @@ def _init():
     })
 
     # Initialize required core packages. Order is important.
-    autoload = ('theme', 'cron', 'reload', 'update', 'setup', 'cleanup', 'odm_http_api', 'plugman')
+    autoload = ('theme', 'cron', 'stats', 'reload', 'update', 'setup', 'cleanup', 'odm_http_api', 'plugman')
     for pkg_name in autoload:
         import_module('pytsite.' + pkg_name)