+++ pytsite/form/__init__.py
@@ -4,14 +4,13 @@
 from . import _error as error, _cache as cache
 from ._form import Form
 
-
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 
 def _init():
-    from pytsite import assetman, tpl, lang, router, http_api, events
+    from pytsite import assetman, tpl, lang, router, http_api, cron, stats
     from . import _http_api
 
     lang.register_package(__name__)
@@ -23,7 +22,11 @@ def _init():
     http_api.handle('POST', 'form/widgets/<uid>', _http_api.get_widgets, 'pytsite.form@post_get_widgets')
     http_api.handle('POST', 'form/validate/<uid>', _http_api.post_validate, 'pytsite.form@post_validate')
 
-    events.listen('pytsite.cron.1min', cache.cleanup)
+    # Cron tasks
+    cron.every_min(cache.cleanup)
+
+    # Stats update
+    stats.on_update(lambda: 'Forms cache size: {}'.format(cache.get_size()))
 
 
 _init()