+++ pytsite/geo_ip/__init__.py
@@ -4,19 +4,18 @@
 from . import _model as model, _error as error
 from ._api import resolve
 
-
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 
 def __init():
-    from pytsite import odm, events
+    from pytsite import odm, events, cron
     from . import _eh
 
     odm.register_model('geo_ip', model.GeoIP)
 
-    events.listen('pytsite.cron.weekly', _eh.pytsite_cron_weekly)
+    cron.weekly(_eh.cron_weekly)
     events.listen('pytsite.update', _eh.pytsite_update)